fix: alteracoes importantes para o funcionamento do JWT

+++ KentoCafe/src/main/java/school/sptech/KentoCafe/dto/funcionario/FuncionarioResponse.java
@@ -1,7 +1,7 @@
 package school.sptech.KentoCafe.dto.funcionario;
 
 public class FuncionarioResponse {
-    private Integer id;
+    private Long id;
     private String nome;
     private String senha;
     private String email;
@@ -10,7 +10,7 @@ public class FuncionarioResponse {
     public FuncionarioResponse() {
     }
 
-    public FuncionarioResponse(Integer id, String nome, String senha, String email, Boolean gerente) {
+    public FuncionarioResponse(Long id, String nome, String senha, String email, Boolean gerente) {
         this.id = id;
         this.nome = nome;
         this.senha = senha;
@@ -18,11 +18,11 @@ public class FuncionarioResponse {
         this.gerente = gerente;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
